+++ CHANGELOG.md
@@ -11,9 +11,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [2.0.1] (unreleased)
 
+### Added
+- Add clean_forgotten_scans(). [#171](https://github.com/greenbone/ospd/pull/171)
+- Extend OSP with finished_hosts to improve resume task.  [#177](https://github.com/greenbone/ospd/pull/177)
+
 ### Changed
 - Set loglevel to debug for some message. [#159](https://github.com/greenbone/ospd/pull/159)
 - Improve error handling when stop a scan. [#163](https://github.com/greenbone/ospd/pull/163)
+- Check the existence and status of an scan_id. [#179](https://github.com/greenbone/ospd/pull/179)
 
 ### Fixed
 - Fix set permission in unix socket. [#157](https://github.com/greenbone/ospd/pull/157)